requirements.txt
Jinja2==2.10
colorlog==4.0.2
hvac==0.7.2
toolz==0.9.0
PyYAML==5.1
docopt==0.6.2
pathspec==0.8.0
